Some changes in the twirly examples to be analyzed by forest

+++ test/commutativity/twirly/cross_iteration_deps/cross_iteration_deps.c
@@ -1,8 +1,8 @@
 int main() {
   unsigned int array[] = {3, 5, 15, 99, 11, 33, 5, 67};
-  const size_t N = sizeof(array) / sizeof(unsigned int);
+  const int N = sizeof(array) / sizeof(unsigned int);
 
-  for (size_t i = 1; i < N; ++i)
+  for (int i = 1; i < N; ++i)
     array[i] += array[i - 1];
 
   return 0;
