+++ test/commutativity/twirly/cross_iteration_deps_fake/cross_iteration_deps_fake.c
@@ -2,7 +2,7 @@ int main() {
   unsigned int array[] = {0};
   const int N = sizeof(array) / sizeof(unsigned int);
 
-  for (size_t i = 1; i < N; ++i)
+  for (int i = 1; i < N; ++i)
     array[i] += array[i - 1];
 
   return 0;
